requirements.txt
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
future==0.16.0
gym==0.10.3
idna==2.6
kiwisolver==1.0.1
matplotlib==2.2.0
numpy==1.14.1
pyglet==1.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
urllib3==1.22
blackhc.mdp==1.0.6
